Move basic styles to their own plugin
