+++ addon/model/util/constants.ts
@@ -1,12 +1,10 @@
-import {
-  boldMarkSpec,
-  highlightMarkSpec,
-  italicMarkSpec,
-  strikethroughMarkSpec,
-  underlineMarkSpec,
-} from '@lblod/ember-rdfa-editor/model/mark';
+import { highlightMarkSpec } from '@lblod/ember-rdfa-editor/model/mark';
 import { TextAttribute } from '@lblod/ember-rdfa-editor/commands/text-properties/set-property-command';
 import { SpecAttributes } from '@lblod/ember-rdfa-editor/model/marks-registry';
+import { boldMarkSpec } from '@lblod/ember-rdfa-editor/plugins/basic-styles/marks/bold';
+import { italicMarkSpec } from '@lblod/ember-rdfa-editor/plugins/basic-styles/marks/italic';
+import { underlineMarkSpec } from '@lblod/ember-rdfa-editor/plugins/basic-styles/marks/underline';
+import { strikethroughMarkSpec } from '@lblod/ember-rdfa-editor/plugins/basic-styles/marks/strikethrough';
 
 export const NON_BLOCK_NODES = new Set(['b', 'strong', 'i', 'em', 'span', 'a']);
 export const LIST_TYPES = new Set(['li', 'ul', 'ol']);
