--- lib/dummy-plugin/app/dummy-plugin.js
@@ -1,1 +1,0 @@
-export { default } from 'dummy-plugin/dummy-plugin';
