--- lib/dummy-plugin/app/initializers/dummy-plugin.js
@@ -1,1 +1,0 @@
-export { default } from 'dummy-plugin/initializers/dummy-plugin';
