--- lib/dummy-plugin/index.js
@@ -1,10 +1,0 @@
-/* eslint-env node */
-'use strict';
-
-module.exports = {
-  name: 'dummy-plugin',
-
-  isDevelopingAddon() {
-    return true;
-  },
-};
