--- lib/dummy-plugin/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "dummy-plugin",
-  "keywords": [
-    "ember-addon"
-  ],
-  "dependencies": {
-    "ember-cli-babel": "*",
-    "ember-cli-typescript": "*"
-  }
-}
