+++ package.json
@@ -99,7 +99,7 @@
     "@types/ember__test-helpers": "^2.6.0",
     "@types/ember__utils": "^3.16.2",
     "@types/htmlbars-inline-precompile": "^1.0.1",
-    "@types/qunit": "^2.11.1",
+    "@types/qunit": "^2.11.3",
     "@types/rsvp": "^4.0.3",
     "@types/sinon": "^10.0.6",
     "@typescript-eslint/eslint-plugin": "^5.5.0",
@@ -181,10 +181,7 @@
     "edition": "octane"
   },
   "ember-addon": {
-    "configPath": "tests/dummy/config",
-    "paths": [
-      "lib/dummy-plugin"
-    ]
+    "configPath": "tests/dummy/config"
   },
   "browserslist": [
     "last 1 Chrome version",
