+++ tests/dummy/app/testplugin/dummy-plugin.ts
@@ -1,6 +1,5 @@
 import { EditorPlugin } from '@lblod/ember-rdfa-editor/utils/editor-plugin';
 import Controller from '@lblod/ember-rdfa-editor/model/controller';
-import { ContentChangedEvent } from '@lblod/ember-rdfa-editor/utils/editor-event';
 import GenTreeWalker from '@lblod/ember-rdfa-editor/model/util/gen-tree-walker';
 import ModelNode from '@lblod/ember-rdfa-editor/model/model-node';
 import { toFilterSkipFalse } from '@lblod/ember-rdfa-editor/model/util/model-tree-walker';
@@ -10,8 +9,7 @@ export default class DummyPlugin implements EditorPlugin {
 
   async initialize(controller: Controller): Promise<void> {
     this.controller = controller;
-    this.controller.onEvent('contentChanged', (event: ContentChangedEvent) => {
-      console.log('DUMMY HANDLER');
+    this.controller.onEvent('contentChanged', () => {
       for (const mark of this.controller.ownMarks) {
         this.controller.executeCommand('remove-mark', mark);
       }
