+++ tests/unit/model/model-text-test.ts
@@ -1,6 +1,7 @@
 import { module, test } from 'qunit';
 import ModelText from '@lblod/ember-rdfa-editor/model/model-text';
-import { boldMarkSpec, Mark } from '@lblod/ember-rdfa-editor/model/mark';
+import { Mark } from '@lblod/ember-rdfa-editor/model/mark';
+import { boldMarkSpec } from '@lblod/ember-rdfa-editor/plugins/basic-styles/marks/bold';
 
 module('Unit | model | model-text-test', () => {
   module('Unit | model | model-text-test | sameAs', () => {
