+++ tsconfig.json
@@ -35,8 +35,7 @@
         "tests/dummy/app/*",
         "app/*",
         "lib/plugin-system-dispatcher/app/*",
-        "lib/rdfa-parser/app/*",
-        "lib/dummy-plugin/app/*"
+        "lib/rdfa-parser/app/*"
       ],
       "@lblod/ember-rdfa-editor": [
         "addon"
@@ -70,18 +69,6 @@
       ],
       "rdfa-parser/test-support/*": [
         "lib/rdfa-parser/addon-test-support/*"
-      ],
-      "dummy-plugin": [
-        "lib/dummy-plugin/addon"
-      ],
-      "dummy-plugin/*": [
-        "lib/dummy-plugin/addon/*"
-      ],
-      "dummy-plugin/test-support": [
-        "lib/dummy-plugin/addon-test-support"
-      ],
-      "dummy-plugin/test-support/*": [
-        "lib/dummy-plugin/addon-test-support/*"
       ]
     }
   },
@@ -90,7 +77,6 @@
     "addon/**/*",
     "tests/**/*",
     "types/**/*",
-    "lib/plugin-system-dispatcher/**/*",
     "test-support/**/*",
     "addon-test-support/**/*",
     "lib/dummy-plugin"
